move oai resync to service
