+++ src/published-data/published-data.controller.ts
@@ -48,7 +48,6 @@ import { FilterQuery, QueryOptions } from "mongoose";
 import { DatasetsService } from "src/datasets/datasets.service";
 import { ProposalsService } from "src/proposals/proposals.service";
 import { AttachmentsService } from "src/attachments/attachments.service";
-import { existsSync, readFileSync } from "fs";
 import { HttpService } from "@nestjs/axios";
 import { ConfigService } from "@nestjs/config";
 import { firstValueFrom } from "rxjs";
@@ -59,8 +58,6 @@ import { DatasetClass } from "src/datasets/schemas/dataset.schema";
 @ApiTags("published data")
 @Controller("publisheddata")
 export class PublishedDataController {
-  private doiConfigPath = "./src/config/doiconfig.local.json";
-
   constructor(
     private readonly attachmentsService: AttachmentsService,
     private readonly configService: ConfigService,
@@ -445,7 +442,7 @@ export class PublishedDataController {
 
     let returnValue = null;
     if(OAIServerUri) {
-      returnValue = await this.resyncOAIPublication(id, publishedData, OAIServerUri)
+      returnValue = await this.publishedDataService.resyncOAIPublication(id, publishedData, OAIServerUri);
     }
 
     try {
@@ -461,49 +458,7 @@ export class PublishedDataController {
     return returnValue;
   }
 
-  private async resyncOAIPublication(
-    id: string,
-    publishedData: UpdatePublishedDataDto,
-    OAIServerUri: string
-  ): Promise<IRegister | null> {
-    let doiProviderCredentials = {
-      username: "removed",
-      password: "removed",
-    };
-
-    if (existsSync(this.doiConfigPath)) {
-      doiProviderCredentials = JSON.parse(
-        readFileSync(this.doiConfigPath).toString(),
-      );
-    }
-
-    const resyncOAIPublication = {
-      method: "PUT",
-      body: publishedData,
-      json: true,
-      uri: OAIServerUri + "/" + encodeURIComponent(encodeURIComponent(id)),
-      headers: {
-        "content-type": "application/json;charset=UTF-8",
-      },
-      auth: doiProviderCredentials,
-    };
-
-    try {
-      const res = await firstValueFrom(
-        this.httpService.request({
-          ...resyncOAIPublication,
-          method: "PUT",
-        }),
-      );
-      return res ? res.data : null;
-    } catch (error:any) {
-      handleAxiosRequestError(error, "PublishedDataController.resync");
-      throw new HttpException(
-        `Error occurred: ${error}`,
-        error.response.status || HttpStatus.FAILED_DEPENDENCY,
-      );
-    }
-  }
+  
 }
 
 function formRegistrationXML(publishedData: PublishedData): string {
